% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelplots.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot gene set enrichments as a panelplot}
\usage{
plot_gsea(df, es_col = "E", pval_col = "adj.P.Val", cleanup_names = TRUE)
}
\arguments{
\item{df}{data frame containing the gene set enrichments}

\item{es_col}{the column name with the effect sizes}

\item{pval_col}{column name with the (adjusted) p-values}

\item{cleanup_names}{if TRUE, the module names will be edited to make them more
legible}
}
\value{
A ggplot object.
}
\description{
Generate a ggplot panelplot from a data frame containing gene set enrichments.
}
\details{
And here the gory details.
}
\examples{
# no example yet. soon!
}
